requirements.txt
csv
pandas
mysql-connector-python
scikit-learn
joblib
schedule
requests